Add files via upload
This is going to be my personal blog. I am going to fill in each of the thumbnail pages with an article about something I have mastered as it relates to cybersecurity. I love the look of the swiping, and choose to keep the page simple. Fun fact, my daughter drew the logo for me.